+++ src/backend/core/services/search.py
@@ -253,19 +253,6 @@ def get_rescore(nb_results):
     ]
 
 
-def get_sort(query_keys, order_by, order_direction):
-    """Build OpenSearch sort clause"""
-    # Add sorting logic based on relevance or specified field
-    if "hybrid" in query_keys:
-        # sorting by other field than "_score" is not supported in hybrid search
-        # see: https://github.com/opensearch-project/neural-search/issues/866
-        return {"_score": {"order": order_direction}}
-    if order_by == enums.RELEVANCE:
-        return {"_score": {"order": order_direction}}
-
-    return {order_by: {"order": order_direction}}
-
-
 def get_params(query_keys):
     """Build OpenSearch search parameters"""
     if "hybrid" in query_keys: